requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.7
autopep8==1.4.4
bcrypt==3.1.7
blinker==1.4
boto==2.48.0
boto3==1.4.7
botocore==1.7.32
Brlapi==0.7.0
Brotli==1.0.9
cachetools==4.0.0
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.39.1
django-cors-headers==3.7.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-storages==1.11.1
docutils==0.14
duplicity==0.8.12.0
entrypoints==0.3
environment==1.0.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gitdb2==2.0.3
GitPython==2.1.7
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.7.1
httplib2==0.14.0
idna==2.6
importlib-metadata==1.5.0
jmespath==0.9.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.0.3
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.0.1
olefile==0.44
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
Pillow==6.2.2
protobuf==3.6.1
psycopg2-binary==2.8.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.6.1
python-debian===0.1.36ubuntu1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
pyuca==1.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.2.0
rsa==4.0
s3transfer==0.1.11
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.3.12
sqlparse==0.3.1
stripe==2.56.0
systemd-python==234
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.24.2
usb-creator==0.3.7
virtualenv==20.4.6
wadllib==1.3.3
whitenoise==5.2.0
xkit==0.0.0
zipp==1.0.0
